Introduce db connection pool
Using a connection pool makes sure that a fixed set of open db
connection are available as long the applicaiton runs. Each db
connection is obtained from the pool and returned in a trycatch
finally statement. When the application stops the ppol gets
closed making sure no leaking connections to be around after
the application has exited.