Search for better failure indicator
Grep'ing for "Exception" is not sufficient for using it as failure
indicator (some log output includes Exception classes).

As `ERROR` does not seem to fit as well (there are error logs
which should be just warnings), we will search for the feeder
summary (though, it needs an INFO log level at least).