Make data import work on Windows platform
Checking for running feeder task is quite fragile under Windows
as there is no real equivalent for `ps aux`.

Currently, there is a loop which checks if a java.exe process
(at best the feeder process) vanishes from the tasklist. However,
this is rather buggy, when another java process gets started or
stopped during the import process. This may lead to an endless
loop in worst case.