Finish document upload feature
Completes feature by adding capabilities to

  - update existing document
  - handle orphaned document (w/o database entry)
  - handle orphaned database entry (w/o document)
  - improve validation and messaging
  - upload literature documents

Literature is now stored in an "Lt" folder which can be
accessed by anyone so no country code is needed here.

For this reason the former `country_code` column changed to
`directory` which, however, still accepts  two chars only which
indicates to be the country code as discriminator. `Lt` stands
for literature documents.