Add hint combinations are created during import process
The user may not be aware, that combinations of parameter, PNS and
topic only can be created when measurements are imported. The fact
that Probe does not have a reference to the actual parameter being
measured makes this neccessary.

However, a hint saying that import process is creating the required
combinations shall make it clear to the user at this point.